Reorganize (#2539)
* remove unused files

* remove functions

* add back for sls

* reorganize

* fix paths
