+++ services/app-api/handlers/measures/tests/get.test.ts
@@ -1,12 +1,12 @@
 import { getMeasure, listMeasures } from "../get";
 
-import dbLib from "../../../libs/dynamodb-lib";
+import dbLib from "../../../utils/dynamo/dynamodb-lib";
 
 import { APIGatewayProxyEvent } from "aws-lambda";
-import { testEvent } from "../../../test-util/testEvents";
-import { convertToDynamoExpression } from "../../dynamoUtils/convertToDynamoExpressionVars";
+import { proxyEvent } from "../../../utils/testing/proxyEvent";
+import { convertToDynamoExpression } from "../../../utils/dynamo/convertToDynamoExpressionVars";
 
-jest.mock("../../../libs/dynamodb-lib", () => ({
+jest.mock("../../../utils/dynamo/dynamodb-lib", () => ({
   __esModule: true,
   default: {
     get: jest.fn().mockReturnValue("single measure"),
@@ -14,12 +14,12 @@ jest.mock("../../../libs/dynamodb-lib", () => ({
   },
 }));
 
-jest.mock("../../../libs/authorization", () => ({
+jest.mock("../../../utils/auth/authorization", () => ({
   __esModule: true,
   isAuthorized: jest.fn().mockReturnValue(true),
 }));
 
-jest.mock("../../../libs/debug-lib", () => ({
+jest.mock("../../../utils/debugging/debug-lib", () => ({
   __esModule: true,
   init: jest.fn(),
   flush: jest.fn(),
@@ -30,7 +30,7 @@ jest.mock("../../dynamoUtils/createCompoundKey", () => ({
   createCompoundKey: jest.fn().mockReturnValue("FL2020ACSFUA-AD"),
 }));
 
-jest.mock("../../dynamoUtils/convertToDynamoExpressionVars", () => ({
+jest.mock("../../../utils/dynamo/convertToDynamoExpressionVars", () => ({
   __esModule: true,
   convertToDynamoExpression: jest.fn().mockReturnValue({ testValue: "test" }),
 }));
@@ -38,7 +38,7 @@ jest.mock("../../dynamoUtils/convertToDynamoExpressionVars", () => ({
 describe("Test Get Measure Handlers", () => {
   test("Test Fetching a Measure", async () => {
     const event: APIGatewayProxyEvent = {
-      ...testEvent,
+      ...proxyEvent,
       body: `{"data": {}, "description": "sample desc"}`,
       headers: { "cognito-identity-id": "test" },
       pathParameters: { coreSet: "ACS" },
@@ -60,7 +60,7 @@ describe("Test Get Measure Handlers", () => {
 
   test("Test Successfully Fetching a List of Measures", async () => {
     const event: APIGatewayProxyEvent = {
-      ...testEvent,
+      ...proxyEvent,
       body: `{"data": {}, "description": "sample desc"}`,
       headers: { "cognito-identity-id": "test" },
       pathParameters: { coreSet: "ACS", state: "FL", year: "2020" },
@@ -81,7 +81,7 @@ describe("Test Get Measure Handlers", () => {
 
   test("Test Fetching a List of Measures with no Path Parameters", async () => {
     const event: APIGatewayProxyEvent = {
-      ...testEvent,
+      ...proxyEvent,
       body: `{"data": {}, "description": "sample desc"}`,
       headers: { "cognito-identity-id": "test" },
       pathParameters: null,
